Refactor dev aggregator and fix cgo
